Move metrics, add none mode to norm, minor fixes
