+++ deliravision/torch/utils/__init__.py
@@ -1,1 +1,1 @@
-from .tensor_ops import make_onehot_torch, make_onehot_npy
+from .tensor_ops import make_onehot_torch
