+++ tests/models/test_decoders_torch.py
@@ -1,1 +1,1 @@
-# ToDO: Decoder tests
+# TODO: Decoder tests
